feat: Added quickind function to get_next_line.h
Optimised linear search for a character in a string. Uses longword to compare up to 8 bytes at a time.
If a character is not found in a string, returns a pointer to the null-byte at the end of the string.
This works specifically in the case of gnl and only on 32-bit and 64-bit longwords, so do not used this
version in a context where portability is important. Portable code which does not follow the42 norm will
be added later to the repo.